mooo

+++ include/display.h
@@ -5,4 +5,4 @@ extern void set_intake_text(bool on, bool inverse);
 // extern void set_control_text(const std::string text);
 extern void add_piston_usage();
 extern void display_tick();
-extern void set_motor_info(bool currentIssue, bool tempIssue, int maxTemp);
+extern void set_motor_info(bool currentIssue, bool tempIssue, double maxTemp);
